requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.4.0
rich==13.7.0
